+++ source/testCases/testLidar.c
@@ -5,13 +5,12 @@
 void testLidarAcquisition(void) {
 
     PolarCoordinate* scan_data; 
-    int point_number;
+    size_t point_number;
 
     if (lidar_start() != 0) {
         printf("Erreur : Impossible de démarrer le Lidar.\n");
         return;
     }
-    
 
     for (int i =0; i<3;i++)
     {
@@ -21,7 +20,7 @@ void testLidarAcquisition(void) {
         scan_data = lidar_update_scan(&point_number);
 
     
-        printf("\n--- AFFICHAGE DES POINTS ACQUIS (%d points) ---\n", point_number);
+        printf("\n--- AFFICHAGE DES POINTS ACQUIS (%ld points) ---\n", point_number);
         printf("Index | Angle (°) | Distance (mm)\n");
         printf("------------------------------------\n");
 
@@ -29,7 +28,6 @@ void testLidarAcquisition(void) {
         {
             printf("[%d] | %.2f° | %.2f mm\n", i, scan_data[i].theta, scan_data[i].dist);
         }
-
         printf("------------------------------------\n");
         
     }
